// components/UserProfile/types.ts
export type TabType =
  | "personal"
  | "gender"
  | "emprendimiento"
  | "ideaNegocio"
  | "innovacion"
  | "mercado"
  | "contabilidadFinanzas"
  | "formalizacion"
  | "financiamiento"
  | "capacitacion";

export type TabData = {
  personal: any;
  gender: any;
  emprendimiento: any;
  ideaNegocio: any;
  innovacion: any;
  mercado: any;
  contabilidadFinanzas: any;
  formalizacion: any;
  financiamiento: any;
  capacitacion: any;
};